@@ -25,43 +25,43 @@ describe ContactTicket do
   end
 
   it "should validate anonymous tickets" do
-    anon_ticket.should be_valid
+    expect(anon_ticket).to be_valid
   end
 
   it "should validate named tickets" do
-    named_ticket.should be_valid
+    expect(named_ticket).to be_valid
   end
 
   it "should return contact error with empty textdetails" do
-    anon_ticket(textdetails: "").should have(1).error_on(:textdetails)
+    expect(anon_ticket(textdetails: "").errors[:textdetails].size).to eq(1)
   end
 
   it "should return contact error with too long name" do
-    named_ticket(name: build_random_string(1251)).should have(1).error_on(:name)
+    expect(named_ticket(name: build_random_string(1251)).errors[:name].size).to eq(1)
   end
 
   it "should return contact error with bad email" do
-    named_ticket(email: build_random_string(12)).should have(1).error_on(:email)
+    expect(named_ticket(email: build_random_string(12)).errors[:email].size).to eq(1)
   end
 
   it "should not be valid if the email contains spaces" do
-    named_ticket(email: "abc @d.com").should have(1).error_on(:email)
+    expect(named_ticket(email: "abc @d.com").errors[:email].size).to eq(1)
   end
 
   it "should not be valid if the email has a dot at the end" do
-    named_ticket(email: "abc@d.com.").should have(1).error_on(:email)
+    expect(named_ticket(email: "abc@d.com.").errors[:email].size).to eq(1)
   end
 
   it "should return contact error with too long email" do
-    named_ticket(email: (build_random_string 1251) + "@a.com").should have(1).error_on(:email)
+    expect(named_ticket(email: (build_random_string 1251) + "@a.com").errors[:email].size).to eq(2)
   end
 
   it "should return contact error with location specific but without link" do
-    anon_ticket(location: "specific").should have(1).error_on(:link)
+    expect(anon_ticket(location: "specific").errors[:link].size).to eq(1)
   end
 
   it "should return contact error with too long textdetails" do
-    anon_ticket(textdetails: build_random_string(1251)).should have(1).error_on(:textdetails)
+    expect(anon_ticket(textdetails: build_random_string(1251)).errors[:textdetails].size).to eq(1)
   end
 
   it "should save the user agent and javascript state" do
@@ -71,16 +71,16 @@ describe ContactTicket do
     ticket = anon_ticket(user_agent: expected_user_agent,
       javascript_enabled: expected_javascript_state)
 
-    ticket.user_agent.should eq expected_user_agent
-    ticket.javascript_enabled.should eq expected_javascript_state
+    expect(ticket.user_agent).to eq expected_user_agent
+    expect(ticket.javascript_enabled).to eq expected_javascript_state
   end
 
   it "should set the javascript state to false by default" do
-    anon_ticket.javascript_enabled.should be_false
+    expect(anon_ticket.javascript_enabled).to be_falsey
   end
 
   it "should make sure that a location is present" do
-    anon_ticket(location: "").should have(1).error_on(:location)
-    anon_ticket(location: nil).should have(1).error_on(:location)
+    expect(anon_ticket(location: "").errors[:location].size).to eq(1)
+    expect(anon_ticket(location: nil).errors[:location].size).to eq(1)
   end
 end
