@@ -7,38 +7,38 @@ describe ReportAProblemTicket do
   end
 
   it "should validate the presence of 'what_wrong'" do
-    ticket(what_wrong: '').should have(1).error_on(:what_wrong)
+    expect(ticket(what_wrong: '').errors[:what_wrong].size).to eq(1)
   end
 
   it "should validate the presence of 'what_doing'" do
-    ticket(what_doing: '').should have(1).error_on(:what_doing)
+    expect(ticket(what_doing: '').errors[:what_doing].size).to eq(1)
   end
 
   it "should filter 'javascript_enabled'" do
-    ticket(javascript_enabled: 'true').javascript_enabled.should be_true
+    expect(ticket(javascript_enabled: 'true').javascript_enabled).to be_truthy
 
-    ticket(javascript_enabled: 'false').javascript_enabled.should be_false
-    ticket(javascript_enabled: 'xxx').javascript_enabled.should be_false
-    ticket(javascript_enabled: '').javascript_enabled.should be_false
+    expect(ticket(javascript_enabled: 'false').javascript_enabled).to be_falsey
+    expect(ticket(javascript_enabled: 'xxx').javascript_enabled).to be_falsey
+    expect(ticket(javascript_enabled: '').javascript_enabled).to be_falsey
   end
 
   it "should filter 'page_owner'" do
-    ticket(page_owner: 'abc').page_owner.should eq('abc')
-    ticket(page_owner: 'number_10').page_owner.should eq('number_10')
+    expect(ticket(page_owner: 'abc').page_owner).to eq('abc')
+    expect(ticket(page_owner: 'number_10').page_owner).to eq('number_10')
 
-    ticket(page_owner: nil).page_owner.should be_nil
-    ticket(page_owner: '').page_owner.should be_nil
-    ticket(page_owner: 'spaces not allowed').page_owner.should be_nil
-    ticket(page_owner: '<hax>').page_owner.should be_nil
-    ticket(page_owner: 'S&P').page_owner.should be_nil
+    expect(ticket(page_owner: nil).page_owner).to be_nil
+    expect(ticket(page_owner: '').page_owner).to be_nil
+    expect(ticket(page_owner: 'spaces not allowed').page_owner).to be_nil
+    expect(ticket(page_owner: '<hax>').page_owner).to be_nil
+    expect(ticket(page_owner: 'S&P').page_owner).to be_nil
   end
 
   it "should filter 'source'" do
-    ticket(source: 'mainstream').source.should eq('mainstream')
-    ticket(source: 'page_not_found').source.should eq('page_not_found')
-    ticket(source: 'inside_government').source.should eq('inside_government')
+    expect(ticket(source: 'mainstream').source).to eq('mainstream')
+    expect(ticket(source: 'page_not_found').source).to eq('page_not_found')
+    expect(ticket(source: 'inside_government').source).to eq('inside_government')
 
-    ticket(source: 'xxx').source.should be_nil
+    expect(ticket(source: 'xxx').source).to be_nil
   end
 
   it "should know if it's spam or not" do
