@@ -5,6 +5,6 @@ describe "Redirecting the root URL" do
   specify "visiting the root URL redirects me to the contact index page", :allow_rescue => true do
     get "/"
 
-    response.should redirect_to("/contact")
+    expect(response).to redirect_to("/contact")
   end
 end
