@@ -4,8 +4,8 @@ module PathHelpers
   def i_should_be_on(path_with_query)
     expected = URI.parse(path_with_query)
     current = URI.parse(current_url)
-    current.path.should == expected.path
-    Rack::Utils.parse_query(current.query).should == Rack::Utils.parse_query(expected.query)
+    expect(current.path).to eq(expected.path)
+    expect(Rack::Utils.parse_query(current.query)).to eq(Rack::Utils.parse_query(expected.query))
   end
 end
 
