@@ -4,7 +4,7 @@ WebMock.disable_net_connect!(:allow_localhost => true)
 
 module WebmockHelper
   def no_web_calls_should_have_been_made
-    WebMock.should_not have_requested(:any, /.*/)
+    expect(WebMock).not_to have_requested(:any, /.*/)
   end
 end
 
