@@ -8,12 +8,7 @@ require "action_mailer/railtie"
 require "sprockets/railtie"
 # require "rails/test_unit/railtie"
 
-if defined?(Bundler)
-  # If you precompile assets before deploying to production, use this line
-  Bundler.require(*Rails.groups(:assets => %w(development test)))
-  # If you want your assets lazily compiled in production, use this line
-  # Bundler.require(:default, :assets, Rails.env)
-end
+Bundler.require(*Rails.groups)
 
 module Feedback
   class Application < Rails::Application
@@ -68,7 +63,7 @@ module Feedback
     # Version of your assets, change this if you want to expire all your assets
     config.assets.version = '1.0'
 
-    config.assets.prefix = 'feedback' # this has to match the path configured in puppet and deploy scripts.
+    config.assets.prefix = '/feedback' # this has to match the path configured in puppet and deploy scripts.
 
     # Precompile additional assets (application.js, application.css, and all non-JS/CSS are already added)
     config.assets.precompile += %w(feedback.js feedback.css feedback-ie6.css)
